api-gateway/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.7.0